+++ src/Sentry.Unity.Native/SentryNative.cs
@@ -87,7 +87,7 @@ public static class SentryNative
         }
         options.CrashedLastRun = () => crashedLastRun;
 
-        if (options.EnableAppHangTracking)
+        if (options.Experimental.EnableNativeAppHangTracking)
         {
             Logger?.LogDebug("Starting the app-hang heartbeat coroutine.");
             SentryMonoBehaviour.Instance.StartAppHangHeartbeat(SentryNativeBridge.AppHangHeartbeat);